// src/app/modules/user/model/meal.ts
export class Meals {
    ID: number;
    Title: any;
    Description: any;
    UserID: number;
    Date: Date;
    Week: Date;
    Start_Weight: any;
    End_Weight: any;
    Note: any;
    Status: any;
    Breakfast1: any;
    Breakfast2: any;
    Breakfast3: any;
    Breakfast4: any;
    Breakfast5: any;
    Breakfast6: any;
    Breakfast7: any;
    Breakfast8: any;
    Breakfast9: any;
    Breakfast10: any;
    Breakfast11: any;
    Breakfast12: any;
    Breakfast13: any;
    Breakfast14: any;
    Breakfast15: any;
    Breakfast16: any;
    Breakfast17: any;
    Breakfast18: any;
    Breakfast19: any;
    Breakfast20: any;
    Breakfast21: any;
    Breakfast22: any;
    Breakfast23: any;
    Breakfast24: any;
    Breakfast25: any;
    Lunch1: any;
    Lunch2: any;
    Lunch3: any;
    Lunch4: any;
    Lunch5: any;
    Lunch6: any;
    Lunch7: any;
    Lunch8: any;
    Lunch9: any;
    Lunch10: any;
    Lunch11: any;
    Lunch12: any;
    Lunch13: any;
    Lunch14: any;
    Lunch15: any;
    Lunch16: any;
    Lunch17: any;
    Lunch18: any;
    Lunch19: any;
    Lunch20: any;
    Lunch21: any;
    Lunch22: any;
    Lunch23: any;
    Lunch24: any;
    Lunch25: any;
    Dinner1: any;
    Dinner2: any;
    Dinner3: any;
    Dinner4: any;
    Dinner5: any;
    Dinner6: any;
    Dinner7: any;
    Dinner8: any;
    Dinner9: any;
    Dinner10: any;
    Dinner11: any;
    Dinner12: any;
    Dinner13: any;
    Dinner14: any;
    Dinner15: any;
    Dinner16: any;
    Dinner17: any;
    Dinner18: any;
    Dinner19: any;
    Dinner20: any;
    Dinner21: any;
    Dinner22: any;
    Dinner23: any;
    Dinner24: any;
    Dinner25: any;
    Created_By: any;
    Created_Date: Date;
    Modified_By: any;
    Modified_Date: Date;
}